refactor(chat): extract ChatInput component and improve focus behavior
- Extract chat input functionality into a new ChatInput.vue component
- Implement two-way binding and validation logic in ChatInput
- Add exposed focus method to ChatInput for external control
- Update index.vue to use new ChatInput component
- Maintain focus behavior after message submission and clear
- Follow Vue 3 composition API best practices with <script setup>
- Use TypeScript for improved type safety
- Keep consistent styling with Nuxt UI components
